backend/requirements.txt
# ── Core ──
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
python-multipart==0.0.20

# ── LLM ──
boto3==1.36.4
anthropic==0.43.0

# ── Embeddings ──
sentence-transformers==3.3.1

# ── Vector Stores ──
chromadb==0.6.3
opensearch-py==2.8.0
requests-aws4auth==1.3.1

# ── Document Processing ──
pypdf==5.1.0
python-docx==1.1.2
beautifulsoup4==4.12.3
lxml==5.3.0
tiktoken==0.8.0
unstructured==0.16.12

# ── Streaming ──
sse-starlette==2.2.1

# ── Utilities ──
tenacity==9.0.0
structlog==24.4.0
httpx==0.28.1

# ── Testing ──
pytest==8.3.4
pytest-asyncio==0.25.0
